% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_dashboard.R
\name{activity_dashboard}
\alias{activity_dashboard}
\title{Activity Dashboard}
\usage{
activity_dashboard(eventlog)
}
\arguments{
\item{eventlog}{Event log to be used for dashboard}
}
\description{
Activity Dashboard
}
