% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_dashboard.R
\name{performance_dashboard}
\alias{performance_dashboard}
\title{Performance Dashboard}
\usage{
performance_dashboard(eventlog)
}
\arguments{
\item{eventlog}{Event log to be used for dashboard}
}
\description{
Performance Dashboard
}
