% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processmonitR.IACSmod.R
\docType{package}
\name{processmonitR}
\alias{processmonitR}
\alias{processmonitR-package}
\title{processmonitR - Building dashboards for process monitoring in R}
\description{
Functions for constructing dashboards for business process monitoring. Building on the event log objects class from package bupaR. Allows the use to assemble custom shiny dashboards based on process data.
}
