% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_dashboard.R
\name{resource_dashboard}
\alias{resource_dashboard}
\title{Resource Dashboard}
\usage{
resource_dashboard(eventlog)
}
\arguments{
\item{eventlog}{Event log to be used for dashboard}
}
\description{
Resource Dashboard
}
