% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rework_dashboard.R
\name{rework_dashboard}
\alias{rework_dashboard}
\title{Rework Dashboard}
\usage{
rework_dashboard(eventlog)
}
\arguments{
\item{eventlog}{Event log to be used for dashboard}
}
\description{
Rework Dashboard
}
